@@ -1,63 +1,7 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux'
-import './App.css';
-import { Route, Switch, Redirect } from 'react-router-dom';
+import Layout from "./components/Layout/Layout";
 
-
-import Homepage from './Pages/homepage/Homepage.component'
-import ShopPage from './Pages/shop/Shop.component'
-import SignInAndSignUpPage from './Pages/sign-in-and-sign-up/sign-in-and-sign-up.component'
-import CheckoutPage from './Pages/checkout/checkout.component'
-
-import Header from './components/header/header.component'
-
-
-import { auth, createUserProfileDocument } from './firebase/firebase.utils'
-import { setCurrentUser } from './redux/user/user.actions'
-import { createStructuredSelector } from 'reselect'
-import { selectCurrentUser } from './redux/user/user.selector'
-
-class App extends Component {
-  unsubscribeFromAuth = null
-  componentDidMount() {
-    const { setCurrentUser } = this.props
-    this.unsubscribeFromAuth = auth.onAuthStateChanged(async userAuth => {
-      if (userAuth) {
-
-        const userRef = await createUserProfileDocument(userAuth)
-
-        userRef.onSnapshot(snapshot => {
-          setCurrentUser({
-            currentUser: {
-              id: snapshot.id,
-              ...snapshot.data()
-            }
-          })
-        })
-      }
-      setCurrentUser(userAuth)
-    })
-  }
-  componentWillUnmount() {
-    this.unsubscribeFromAuth()
-  }
-  render() {
-    return <div className='App'>
-      <Header />
-      <Switch>
-        <Route exact path='/' component={Homepage} />
-        <Route path='/shop' component={ShopPage} />
-        <Route path='/checkout' component={CheckoutPage} />
-        <Route path='/signin' render={() => this.props.currentUser ? (<Redirect to='/' />) : <SignInAndSignUpPage />} />
-      </Switch>
-    </div>;
-  }
+function App() {
+  return <Layout />;
 }
-const mapStateToProps = createStructuredSelector({
-  currentUser: selectCurrentUser,
-})
 
-const mapDispatchToProps = dispatch => ({
-  setCurrentUser: user => dispatch(setCurrentUser(user))
-})
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default App;
